Part : 5 Added Spring security login, signup, refreshtoken

+++ src/main/java/com/codebygaurav/AirbnbApp/service/impl/RoomServiceImpl.java
@@ -3,7 +3,9 @@ package com.codebygaurav.AirbnbApp.service.impl;
 import com.codebygaurav.AirbnbApp.dto.RoomDto;
 import com.codebygaurav.AirbnbApp.entity.Hotel;
 import com.codebygaurav.AirbnbApp.entity.Room;
+import com.codebygaurav.AirbnbApp.entity.User;
 import com.codebygaurav.AirbnbApp.exception.ResourceNotFoundException;
+import com.codebygaurav.AirbnbApp.exception.UnAuthorisedException;
 import com.codebygaurav.AirbnbApp.repository.HotelRepository;
 import com.codebygaurav.AirbnbApp.repository.RoomRepository;
 import com.codebygaurav.AirbnbApp.service.InventoryService;
@@ -11,6 +13,7 @@ import com.codebygaurav.AirbnbApp.service.RoomService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -32,6 +35,12 @@ public class RoomServiceImpl implements RoomService {
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(()-> new ResourceNotFoundException("Hotel not found with ID:"+hotelId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with Id: "+hotelId);
+        }
+
         Room room = modelMapper.map(roomDto, Room.class);
         room.setHotel(hotel);
         room = roomRepository.save(room);
@@ -50,6 +59,12 @@ public class RoomServiceImpl implements RoomService {
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(()-> new ResourceNotFoundException("Hotel not found with ID:"+hotelId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with Id: "+hotelId);
+        }
+
         return hotel.getRooms()
                 .stream()
                 .map((element) -> modelMapper.map(element, RoomDto.class))
@@ -75,6 +90,10 @@ public class RoomServiceImpl implements RoomService {
                 .findById(roomId)
                 .orElseThrow(()-> new ResourceNotFoundException("Hotel not found with ID:"+roomId));
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(room.getHotel().getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with Id: "+roomId);
+        }
         //TODO: delete all future inventory for this room
         inventoryService.deleteAllInventories(room);
         roomRepository.deleteById(roomId);
